trelte: hexedit libGLES_mali.so
The L linker changed dlsym to do a breadth-first search for symbols
rather than looking only in the referenced library.

Our mali blob is linked against libgui.so which is linked against
libEGL.so which causes all kinds of problems because vendor
implementations get resolved to the opengl generic wrappers.

Change the static dependency from libgui.so to liblog.so
(it really already is linked against liblog.so, but linking it
again seemed to do no harm) to workaround this issue.